Fixed MS VS warnings. JIRA CXX-9192
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@78138 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/connect/services/impl/netschedule_api_int.hpp
@@ -51,7 +51,7 @@ void g_AppendClientIPSessionIDHitID(string& cmd, bool = false);
 
 extern NCBI_XCONNECT_EXPORT
 int g_ParseNSOutput(const string& attr_string, const char* const* attr_names,
-        string* attr_values, int attr_count);
+        string* attr_values, size_t attr_count);
 
 void NCBI_XCONNECT_EXPORT NCBI_EntryPoint_xnetscheduleapi(
      CPluginManager<SNetScheduleAPIImpl>::TDriverInfoList&   info_list,
